backend/requirements.txt
django
djangorestframework
django-cors-headers
django-environ
psycopg2-binary
djangorestframework-simplejwt
django-filter
django-channels
channels-redis
celery
redis
django-allauth
dj-rest-auth
python-magic
pdfminer.six
gunicorn
whitenoise
